feat(web): UI improvements - settings redesign, task launcher, sidebar favicons, styling
- Redesign SettingsDialog with simplified structure and AboutTab/DebugSection extraction
- Improve TaskLauncher with time-stable filtering
- Add favicon display in ConversationListItem
- Sync task messages on completion for immediate favicon display
- Extract Execution page components (MessageList, DebugPanel, PermissionDialog, ToolProgress, SpinningIcon)
- Add task-utils helper for favicon extraction
- Add tool-mappings constants
- Add Geist font family
- Update globals.css and tailwind config for new design tokens
- Simplify SpeechSettingsForm
- Various component cleanups (named exports, removed dead code)
- Update integration tests for SettingsDialog, Sidebar, TaskLauncher
