+++ apps/web/src/client/components/TaskLauncher/TaskLauncher.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useState, useEffect, useMemo, useCallback } from 'react';
 import { useNavigate } from 'react-router';
 import { motion, AnimatePresence } from 'framer-motion';
@@ -9,43 +7,46 @@ import { useTaskStore } from '@/stores/taskStore';
 import { getAccomplish } from '@/lib/accomplish';
 import { cn } from '@/lib/utils';
 import { springs } from '@/lib/animations';
-import TaskLauncherItem from './TaskLauncherItem';
+import { TaskLauncherItem } from './TaskLauncherItem';
 import { hasAnyReadyProvider } from '@accomplish_ai/agent-core/common';
 import { Input } from '@/components/ui/input';
 
-export default function TaskLauncher() {
+export function TaskLauncher() {
   const navigate = useNavigate();
   const [searchQuery, setSearchQuery] = useState('');
   const [selectedIndex, setSelectedIndex] = useState(0);
 
   const { isLauncherOpen, launcherInitialPrompt, closeLauncher, tasks, startTask } = useTaskStore();
   const accomplish = getAccomplish();
+  const [openedAt, setOpenedAt] = useState(Date.now);
 
   // Filter tasks by search query (title only)
   const filteredTasks = useMemo(() => {
     if (!searchQuery.trim()) {
       // Show last 7 days when no search
-      // eslint-disable-next-line react-hooks/purity
-      const sevenDaysAgo = Date.now() - 7 * 24 * 60 * 60 * 1000;
+      const sevenDaysAgo = openedAt - 7 * 24 * 60 * 60 * 1000;
       return tasks.filter((t) => new Date(t.createdAt).getTime() > sevenDaysAgo);
     }
     const query = searchQuery.toLowerCase();
     return tasks.filter((t) => t.prompt.toLowerCase().includes(query));
-  }, [tasks, searchQuery]);
+  }, [tasks, searchQuery, openedAt]);
 
   // Total items: "New task" + filtered tasks
   const totalItems = 1 + filteredTasks.length;
 
   // Clamp selected index when results change
   useEffect(() => {
+    // eslint-disable-next-line react-hooks/set-state-in-effect -- intentional: sync derived state from prop change
     setSelectedIndex((i) => Math.min(i, Math.max(0, totalItems - 1)));
   }, [totalItems]);
 
   // Reset state when launcher opens, use initial prompt if provided
   useEffect(() => {
     if (isLauncherOpen) {
+      // eslint-disable-next-line react-hooks/set-state-in-effect -- intentional: reset state on open
       setSearchQuery(launcherInitialPrompt || '');
       setSelectedIndex(0);
+      setOpenedAt(Date.now());
     }
   }, [isLauncherOpen, launcherInitialPrompt]);
 
@@ -134,15 +135,17 @@ export default function TaskLauncher() {
                 animate={{ opacity: 1 }}
                 exit={{ opacity: 0 }}
                 transition={{ duration: 0.15 }}
-                className="fixed inset-0 z-50 bg-black/50 backdrop-blur-sm"
+                className="fixed inset-0 z-50 bg-white/60 backdrop-blur-[12px]"
               />
             </DialogPrimitive.Overlay>
 
             {/* Content */}
             <DialogPrimitive.Content
               className="fixed inset-0 z-50 flex items-start justify-center pt-[20vh]"
+              aria-describedby={undefined}
               onKeyDown={handleKeyDown}
             >
+              <DialogPrimitive.Title className="sr-only">Task Launcher</DialogPrimitive.Title>
               <motion.div
                 initial={{ opacity: 0, scale: 0.95, y: -10 }}
                 animate={{ opacity: 1, scale: 1, y: 0 }}
@@ -194,7 +197,7 @@ export default function TaskLauncher() {
                             : 'text-muted-foreground',
                         )}
                       >
-                        — &quot;{searchQuery}&quot;
+                        — &ldquo;{searchQuery}&rdquo;
                       </span>
                     )}
                   </button>
@@ -244,3 +247,5 @@ export default function TaskLauncher() {
     </DialogPrimitive.Root>
   );
 }
+
+export default TaskLauncher;
