+++ apps/web/src/client/components/TaskLauncher/TaskLauncherItem.tsx
@@ -1,8 +1,8 @@
-'use client';
-
+import { useMemo } from 'react';
 import type { Task } from '@accomplish_ai/agent-core/common';
 import { cn } from '@/lib/utils';
-import { Loader2, CheckCircle2, XCircle, AlertCircle } from 'lucide-react';
+import { Loader2 } from 'lucide-react';
+import { STATUS_COLORS, extractDomains } from '@/lib/task-utils';
 
 interface TaskLauncherItemProps {
   task: Task;
@@ -10,55 +10,52 @@ interface TaskLauncherItemProps {
   onClick: () => void;
 }
 
-function formatRelativeDate(dateString: string): string {
-  const date = new Date(dateString);
-  const now = new Date();
-  const diffDays = Math.floor((now.getTime() - date.getTime()) / (1000 * 60 * 60 * 24));
+export function TaskLauncherItem({ task, isSelected, onClick }: TaskLauncherItemProps) {
+  const domains = useMemo(() => extractDomains(task), [task]);
+  const statusColor = STATUS_COLORS[task.status] || 'bg-muted-foreground';
 
-  if (diffDays === 0) return 'Today';
-  if (diffDays === 1) return 'Yesterday';
-  if (diffDays < 7) {
-    return date.toLocaleDateString('en-US', { weekday: 'long' });
-  }
-  return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric' });
-}
-
-function getStatusIcon(status: Task['status']) {
-  switch (status) {
-    case 'running':
-      return <Loader2 className="h-3 w-3 animate-spin text-primary shrink-0" />;
-    case 'completed':
-      return <CheckCircle2 className="h-3 w-3 text-green-500 shrink-0" />;
-    case 'failed':
-      return <XCircle className="h-3 w-3 text-destructive shrink-0" />;
-    case 'cancelled':
-    case 'interrupted':
-      return <AlertCircle className="h-3 w-3 text-yellow-500 shrink-0" />;
-    default:
-      return null;
-  }
-}
-
-export default function TaskLauncherItem({ task, isSelected, onClick }: TaskLauncherItemProps) {
   return (
     <button
       onClick={onClick}
       className={cn(
-        'w-full text-left px-3 py-2 rounded-md text-sm transition-colors duration-100',
-        'flex items-center gap-2',
+        'w-full text-left p-3 rounded-lg text-xs font-medium transition-colors duration-100',
+        'flex items-center gap-3',
         isSelected ? 'bg-primary text-primary-foreground' : 'text-foreground hover:bg-accent',
       )}
     >
-      {getStatusIcon(task.status)}
-      <span className="truncate flex-1">{task.prompt}</span>
-      <span
-        className={cn(
-          'text-xs shrink-0',
-          isSelected ? 'text-primary-foreground/70' : 'text-muted-foreground',
+      <span className="flex items-center justify-center shrink-0 w-3 h-3">
+        {task.status === 'running' || task.status === 'waiting_permission' ? (
+          <Loader2 className="w-3 h-3 animate-spin text-muted-foreground" />
+        ) : (
+          <span className={cn('w-2 h-2 rounded-full', statusColor)} />
         )}
-      >
-        {formatRelativeDate(task.createdAt)}
       </span>
+      <span className="truncate flex-1 tracking-[0.18px]">{task.prompt}</span>
+      {domains.length > 0 && (
+        <span className={cn('flex items-center shrink-0', domains.length > 1 && 'pr-1')}>
+          {domains.map((domain, i) => (
+            <span
+              key={domain}
+              className={cn(
+                'flex items-center p-0.5 rounded-full bg-white shrink-0 relative',
+                i > 0 && '-ml-1',
+                i === 0 && 'z-30',
+                i === 1 && 'z-20',
+                i === 2 && 'z-10',
+              )}
+            >
+              <img
+                src={`https://www.google.com/s2/favicons?domain=${domain}&sz=16`}
+                alt={domain}
+                className="w-3 h-3 rounded-full"
+                loading="lazy"
+              />
+            </span>
+          ))}
+        </span>
+      )}
     </button>
   );
 }
+
+export default TaskLauncherItem;
