+++ apps/web/src/client/components/execution/SpinningIcon.tsx
@@ -1,0 +1,6 @@
+import { cn } from '@/lib/utils';
+import loadingSymbol from '/assets/loading-symbol.svg';
+
+export const SpinningIcon = ({ className }: { className?: string }) => (
+  <img src={loadingSymbol} alt="" className={cn('animate-spin-ccw', className)} />
+);
