+++ apps/web/src/client/components/layout/ConversationListItem.tsx
@@ -1,20 +1,21 @@
-'use client';
-
+import { useMemo } from 'react';
 import { useNavigate, useLocation } from 'react-router';
 import type { Task } from '@accomplish_ai/agent-core/common';
 import { cn } from '@/lib/utils';
-import { Loader2, CheckCircle2, XCircle, Clock, Square, PauseCircle, X } from 'lucide-react';
+import { X, Loader2 } from 'lucide-react';
 import { useTaskStore } from '@/stores/taskStore';
+import { STATUS_COLORS, extractDomains } from '@/lib/task-utils';
 
 interface ConversationListItemProps {
   task: Task;
 }
 
-export default function ConversationListItem({ task }: ConversationListItemProps) {
+export function ConversationListItem({ task }: ConversationListItemProps) {
   const navigate = useNavigate();
   const location = useLocation();
   const isActive = location.pathname === `/execution/${task.id}`;
   const deleteTask = useTaskStore((state) => state.deleteTask);
+  const domains = useMemo(() => extractDomains(task), [task]);
 
   const handleClick = () => {
     navigate(`/execution/${task.id}`);
@@ -29,30 +30,12 @@ export default function ConversationListItem({ task }: ConversationListItemProps
 
     await deleteTask(task.id);
 
-    // Navigate to home if deleting the currently active task
     if (isActive) {
       navigate('/');
     }
   };
 
-  const getStatusIcon = () => {
-    switch (task.status) {
-      case 'running':
-        return <Loader2 className="h-3 w-3 animate-spin-ccw text-primary shrink-0" />;
-      case 'completed':
-        return <CheckCircle2 className="h-3 w-3 text-green-500 shrink-0" />;
-      case 'failed':
-        return <XCircle className="h-3 w-3 text-red-500 shrink-0" />;
-      case 'cancelled':
-        return <Square className="h-3 w-3 text-zinc-400 shrink-0" />;
-      case 'interrupted':
-        return <PauseCircle className="h-3 w-3 text-amber-500 shrink-0" />;
-      case 'queued':
-        return <Clock className="h-3 w-3 text-amber-500 shrink-0" />;
-      default:
-        return null;
-    }
-  };
+  const statusColor = STATUS_COLORS[task.status] || 'bg-muted-foreground';
 
   return (
     <div
@@ -67,26 +50,61 @@ export default function ConversationListItem({ task }: ConversationListItemProps
       }}
       title={task.summary || task.prompt}
       className={cn(
-        'w-full text-left px-3 py-2 rounded-md text-sm transition-colors duration-200',
-        'text-foreground hover:bg-accent hover:text-accent-foreground',
-        'flex items-center gap-2 group relative cursor-pointer',
-        isActive && 'bg-accent text-accent-foreground',
+        'w-full text-left p-2 rounded-lg text-xs font-medium transition-colors duration-200',
+        'text-foreground hover:bg-[#E8E8E8] hover:text-foreground',
+        'flex items-center gap-3 group relative cursor-pointer',
+        isActive && 'bg-[#EDEBE7] text-foreground',
       )}
     >
-      {getStatusIcon()}
-      <span className="block truncate flex-1">{task.summary || task.prompt}</span>
-      <button
-        onClick={handleDelete}
-        className={cn(
-          'opacity-0 group-hover:opacity-100 transition-opacity duration-200',
-          'p-1 rounded hover:bg-red-100 dark:hover:bg-red-900/20',
-          'text-zinc-400 hover:text-red-600 dark:hover:text-red-400',
-          'shrink-0',
+      <span className="flex items-center justify-center shrink-0 w-3 h-3">
+        {task.status === 'running' || task.status === 'waiting_permission' ? (
+          <Loader2 className="w-3 h-3 animate-spin text-muted-foreground" />
+        ) : (
+          <span className={cn('w-2 h-2 rounded-full', statusColor)} />
         )}
-        aria-label="Delete task"
-      >
-        <X className="h-3 w-3" />
-      </button>
+      </span>
+      <span className="block truncate flex-1 tracking-[0.18px]">{task.summary || task.prompt}</span>
+      <span className="relative flex items-center shrink-0 h-5">
+        {domains.length > 0 && (
+          <span className="flex items-center group-hover:opacity-0 transition-opacity duration-200">
+            {domains.map((domain, i) => (
+              <span
+                key={domain}
+                className={cn(
+                  'flex items-center p-0.5 rounded-full bg-white shrink-0 relative',
+                  i > 0 && '-ml-1',
+                  i === 0 && 'z-30',
+                  i === 1 && 'z-20',
+                  i === 2 && 'z-10',
+                )}
+              >
+                <img
+                  src={`https://www.google.com/s2/favicons?domain=${domain}&sz=16`}
+                  alt={domain}
+                  className="w-3 h-3 rounded-full"
+                  loading="lazy"
+                />
+              </span>
+            ))}
+          </span>
+        )}
+        <button
+          onClick={handleDelete}
+          title="Remove task"
+          className={cn(
+            'absolute right-0 top-1/2 -translate-y-1/2',
+            'opacity-0 pointer-events-none group-hover:opacity-100 group-hover:pointer-events-auto',
+            'transition-opacity duration-200',
+            'p-1 rounded hover:bg-red-100 dark:hover:bg-red-900/20',
+            'text-zinc-400 hover:text-red-600 dark:hover:text-red-400',
+          )}
+          aria-label="Remove task"
+        >
+          <X className="h-3 w-3" />
+        </button>
+      </span>
     </div>
   );
 }
+
+export default ConversationListItem;
