+++ apps/web/src/client/components/settings/SpeechSettingsForm.tsx
@@ -1,29 +1,10 @@
-/**
- * Speech-to-Text Settings Component
- *
- * Allows users to:
- * - Enable/disable speech input
- * - Configure ElevenLabs API key
- * - Verify configuration by attempting transcription
- */
-
 import { useState, useEffect } from 'react';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
-import { Alert, AlertDescription } from '@/components/ui/alert';
 import { Mic, CheckCircle2, AlertCircle, Loader2 } from 'lucide-react';
 import { getAccomplish } from '../../lib/accomplish';
 
 interface SpeechSettingsFormProps {
-  /**
-   * Callback when API key is saved
-   */
   onSave?: () => void;
-
-  /**
-   * Callback when configuration changes
-   */
   onChange?: (config: { apiKey: string; enabled: boolean }) => void;
 }
 
@@ -33,7 +14,6 @@ export function SpeechSettingsForm({ onSave, onChange }: SpeechSettingsFormProps
   const [apiKey, setApiKey] = useState('');
   const [isConfigured, setIsConfigured] = useState(false);
 
-  // Load existing configuration on mount
   useEffect(() => {
     accomplish.speechGetConfig().then((config) => {
       setIsConfigured(config.hasApiKey);
@@ -52,11 +32,10 @@ export function SpeechSettingsForm({ onSave, onChange }: SpeechSettingsFormProps
     setSaveResult(null);
 
     try {
-      // Save the API key
       await accomplish.addApiKey('elevenlabs', apiKey, 'ElevenLabs Speech-to-Text');
       setSaveResult({ success: true, message: 'API key saved successfully' });
       setIsConfigured(true);
-      setApiKey(''); // Clear the input after saving
+      setApiKey('');
       onChange?.({ apiKey, enabled: true });
       onSave?.();
     } catch (error) {
@@ -67,134 +46,103 @@ export function SpeechSettingsForm({ onSave, onChange }: SpeechSettingsFormProps
     }
   };
 
-  const handleClearApiKey = async () => {
-    try {
-      await accomplish.removeApiKey('local-elevenlabs');
-      setApiKey('');
-      setIsConfigured(false);
-      setSaveResult(null);
-      onChange?.({ apiKey: '', enabled: false });
-    } catch (error) {
-      console.error('Failed to remove API key:', error);
-    }
-  };
-
   return (
-    <Card>
-      <CardHeader>
-        <div className="flex items-center gap-2">
-          <Mic className="h-5 w-5 text-blue-500" />
-          <div>
-            <CardTitle>Speech-to-Text</CardTitle>
-            <CardDescription>
-              Enable voice input using ElevenLabs Speech-to-Text API
-            </CardDescription>
-          </div>
+    <div className="group rounded-xl border border-border bg-card p-3.5 transition-all duration-200 hover:border-primary hover:shadow-md">
+      {/* Header: Title + Toggle */}
+      <div className="mb-1.5">
+        <span className="flex items-center gap-1.5 text-[13px] font-semibold text-foreground">
+          <Mic className="h-3.5 w-3.5 text-blue-500" />
+          Speech-to-Text
+        </span>
+      </div>
+
+      {/* Description */}
+      <p className="mb-2.5 text-[11px] leading-relaxed text-muted-foreground">
+        Enable voice input using ElevenLabs Speech-to-Text API
+      </p>
+
+      {/* Info section */}
+      <div className="mb-2.5 flex items-start gap-1.5 rounded-md border border-border bg-background px-2.5 py-2 text-[11px] text-foreground">
+        <AlertCircle className="mt-0.5 h-3.5 w-3.5 shrink-0" />
+        <span>
+          To use speech input, you need an ElevenLabs API key. Get one at{' '}
+          <a
+            href="https://elevenlabs.io/app/settings/api-keys"
+            target="_blank"
+            rel="noopener noreferrer"
+            className="text-blue-500 hover:underline"
+          >
+            elevenlabs.io
+          </a>
+        </span>
+      </div>
+
+      {/* Configured status */}
+      {isConfigured && !apiKey && (
+        <div className="mb-2.5 flex items-center gap-1.5 rounded-md border border-border bg-background px-2.5 py-2 text-[11px] text-foreground">
+          <CheckCircle2 className="h-3.5 w-3.5 shrink-0 text-green-500" />
+          <span>ElevenLabs API key is configured. Enter a new key below to replace it.</span>
         </div>
-      </CardHeader>
+      )}
 
-      <CardContent className="space-y-4">
-        {/* Info section */}
-        <Alert>
-          <AlertCircle className="h-4 w-4" />
-          <AlertDescription className="text-xs">
-            To use speech input, you need an ElevenLabs API key. Get one at{' '}
-            <a
-              href="https://elevenlabs.io/app/settings/api-keys"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-blue-500 hover:underline"
-            >
-              elevenlabs.io
-            </a>
-          </AlertDescription>
-        </Alert>
-
-        {/* Existing configuration status */}
-        {isConfigured && !apiKey && (
-          <Alert>
-            <CheckCircle2 className="h-4 w-4 text-green-500" />
-            <AlertDescription className="text-xs">
-              ElevenLabs API key is configured. Enter a new key below to replace it.
-            </AlertDescription>
-          </Alert>
-        )}
-
-        {/* API Key Input */}
-        <div className="space-y-2">
-          <label className="text-sm font-medium">ElevenLabs API Key</label>
-          <div className="flex gap-2">
-            <Input
-              type="password"
-              placeholder={isConfigured ? '••••••••••••••••' : 'xi-...'}
-              value={apiKey}
-              onChange={(e) => {
-                setApiKey(e.target.value);
-                setSaveResult(null);
-              }}
-              disabled={isLoading}
-            />
-            {(apiKey || isConfigured) && (
-              <Button
-                type="button"
-                onClick={handleClearApiKey}
-                variant="ghost"
-                size="icon"
-                disabled={isLoading}
-              >
-                <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    strokeWidth={2}
-                    d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
-                  />
-                </svg>
-              </Button>
-            )}
-          </div>
-          <p className="text-xs text-muted-foreground">
-            Your API key is stored securely in your system keychain.
-          </p>
-        </div>
-
-        {/* Action Buttons */}
-        <div className="flex gap-2">
-          <Button
+      {/* API Key Input */}
+      <div className="space-y-2">
+        <label className="text-[11px] font-medium text-foreground">ElevenLabs API Key</label>
+        <div className="flex gap-1.5">
+          <Input
+            type="password"
+            placeholder={isConfigured ? '••••••••••••••••' : 'xi-...'}
+            value={apiKey}
+            onChange={(e) => {
+              setApiKey(e.target.value);
+              setSaveResult(null);
+            }}
+            disabled={isLoading}
+            className="h-7 text-[11px] px-2"
+          />
+          <button
             onClick={handleSaveApiKey}
             disabled={isLoading || !apiKey.trim()}
-            className="flex-1"
+            className="inline-flex h-7 items-center justify-center rounded-md bg-primary px-3 text-[11px] font-medium text-primary-foreground transition-colors hover:bg-primary/90 disabled:pointer-events-none disabled:opacity-50"
           >
-            {isLoading && <Loader2 className="h-4 w-4 mr-2 animate-spin" />}
-            Save API Key
-          </Button>
+            {isLoading ? <Loader2 className="h-3 w-3 animate-spin" /> : 'Save'}
+          </button>
         </div>
+        <p className="text-[10px] text-muted-foreground">
+          Your API key is stored securely in your system keychain.
+        </p>
+      </div>
 
-        {/* Save Result */}
-        {saveResult && (
-          <Alert variant={saveResult.success ? 'default' : 'destructive'}>
-            {saveResult.success ? (
-              <CheckCircle2 className="h-4 w-4" />
-            ) : (
-              <AlertCircle className="h-4 w-4" />
-            )}
-            <AlertDescription className="text-xs">{saveResult.message}</AlertDescription>
-          </Alert>
-        )}
-
-        {/* Usage Instructions */}
-        <div className="rounded-lg bg-blue-50 dark:bg-blue-950 p-3 text-sm">
-          <p className="font-medium text-blue-900 dark:text-blue-100 mb-2">How to use:</p>
-          <ul className="text-blue-800 dark:text-blue-200 space-y-1 text-xs">
-            <li>
-              • <strong>Click the microphone button</strong> to start recording, click again to stop
-            </li>
-            <li>
-              • <strong>Hold Alt key</strong> to record voice input (push-to-talk mode)
-            </li>
-          </ul>
+      {/* Save Result */}
+      {saveResult && (
+        <div
+          className={`mt-2 flex items-center gap-1.5 rounded-md px-2 py-1.5 text-[10px] ${
+            saveResult.success
+              ? 'bg-green-50 text-green-700 dark:bg-green-950 dark:text-green-300'
+              : 'bg-red-50 text-red-700 dark:bg-red-950 dark:text-red-300'
+          }`}
+        >
+          {saveResult.success ? (
+            <CheckCircle2 className="h-3 w-3" />
+          ) : (
+            <AlertCircle className="h-3 w-3" />
+          )}
+          {saveResult.message}
         </div>
-      </CardContent>
-    </Card>
+      )}
+
+      {/* Usage Instructions */}
+      <div className="mt-2.5 rounded-md bg-blue-50 p-2.5 text-[11px] dark:bg-blue-950">
+        <p className="mb-1.5 font-medium text-blue-900 dark:text-blue-100">How to use:</p>
+        <ul className="space-y-1 text-[10px] text-blue-800 dark:text-blue-200">
+          <li>
+            <strong>Click the microphone button</strong> to start recording, click again to stop
+          </li>
+          <li>
+            <strong>Hold Alt key</strong> to record voice input (push-to-talk mode)
+          </li>
+        </ul>
+      </div>
+    </div>
   );
 }
